Tenant_Id Instead of Tenant_Name
I think instructing users to enter the Tenant ID will be more deterministic. My understanding is that you mean to ingest either the tenant id (listed in each application's app registration page) or one of the tenant domain names. 

The use of "Tenant Name" becomes even more dangerous when you notice that, In Properties section of the Azure AD tenant in the portal, "Name" is an editable field that can include spaces. This field is not useful to this python code.

Finally, the relevant code snip in the docs (https://docs.microsoft.com/en-us/azure/active-directory/develop/quickstart-v2-python-daemon) reference a "Tenant_ID" in the example code snippet (although, the use of "Tenant name" is clarified in the second bullet-point below the code)